move container api testcase to issue
